AudienceReview Class elaboration
AudienceReview class is defined with necessary base attributes (all of which are private access). A constructor method is created to allow the initialization /or creation of a new review.  NOTE: Depending on whether or not it's necessary we could add a method later to display all of the review information, though I am unsure if it would mesh well with the UI. Getters were also made for every attribute.

(Don't mix coffee with energy drinks; My kidneys are paying the price for my hubris)