ReviewForum setup
ReviewForum now handles the creation/initialization of an ArrayList & related methods.
ArrayList will contain every Review in the forum / program.
Method addReview() adds a new review to ArrayList when called; displayReview() prints out review info.

NOTE: Since FX is yet to be implemented (as of this moment) displayReview() defaults to printing in the console. Once JavaFX does get set up on the front end, we need to (maybe) rework displayReview() so it functions specifically with the UI.